Give up wxOwnerDrawnComboBox because it looks different from the native one; use text ctrl instead of combobox for find panel.

+++ src/editor/text_line.cc
@@ -371,7 +371,7 @@ IndentProp TextLine::GetIndentProp() const {
   }
 
   prop.type = data_[0] == kTabChar ? kTabIndent : kSpaceIndent;
-  for (size_t i = 1; i < prop.length; ++i) {
+  for (int i = 1; i < prop.length; ++i) {
     if (data_[i] != data_[0]) {
       prop.type = kMixedIndent;
       break;
